0.5.0-rc1

See merge request nunet/device-management-service!693

+++ db/repositories/utils_test.go
@@ -1,3 +1,11 @@
+// Copyright 2024, Nunet
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and limitations under the License.
+
 package repositories
 
 import (
@@ -14,12 +22,12 @@ func TestUpdateField(t *testing.T) {
 	// Test case: Updating a field in a struct (not a pointer)
 	modified1, err := UpdateField(types.FreeResources{Resources: types.Resources{RAM: types.RAM{Size: 1}}}, "RAM", types.RAM{Size: 2})
 	assert.NoError(t, err)
-	assert.Equal(t, uint64(2), modified1.RAM.Size)
+	assert.Equal(t, float64(2), modified1.RAM.Size)
 
 	// Test case: Updating a field in a struct (using a pointer)
 	modified2, err := UpdateField(&types.FreeResources{Resources: types.Resources{RAM: types.RAM{Size: 1}}}, "RAM", types.RAM{Size: 2})
 	assert.NoError(t, err)
-	assert.Equal(t, uint64(2), modified2.RAM.Size)
+	assert.Equal(t, float64(2), modified2.RAM.Size)
 
 	// Test case: Attempting to update a non-existent field results in an error
 	_, err = UpdateField(types.FreeResources{Resources: types.Resources{RAM: types.RAM{Size: 1}}}, "WRAM", types.RAM{Size: 2})
